+++ test/frontend/async-function-test.ts
@@ -1,0 +1,6 @@
+async function load() {}
+export async function save() {}
+export default async function named() {}
+const expression = async function () {};
+const arrow = async () => {};
+class Worker { async run() {} }